(trunk) Replaced source:/gplates/trunk using source:/gplates/branches/reconstruct-rasters-2010-jun-24@11590 incorporating log:/gplates/branches/reconstruct-rasters-2010-jun-24@10300:11590
This is "merging one branch to another without previous sync". The merge was performed using Tortoise SVN's Merge tool, from[[br]]
source:/gplates/branches/reconstruct-rasters-2010-jun-24@10299 to[[br]]
source:/gplates/branches/reconstruct-rasters-2010-jun-24@11590, [[br]]
from inside the root directory of a working copy of the trunk.

git-svn-id: http://svn.gplates.org/projects/gplates/trunk@11591 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/utils/ReferenceCount.h
@@ -80,7 +80,7 @@ namespace GPlatesUtils
 		/**
 		 * The type used to store the reference-count of an instance of this class.
 		 */
-		typedef long ref_count_type;
+		typedef int ref_count_type;
 
 		/**
 		 * Constructor.
